Rename store services to logic
