+++ packages/services/f-statistics/src/store/logic/startBatchPublishTimer.js
